+++ client/src/__tests__/Signup/Signup.test.tsx
@@ -8,14 +8,12 @@ import { Provider } from "react-redux";
 import { Router } from "react-router-dom";
 import { createMemoryHistory } from "history";
 import { render, cleanup, screen, fireEvent } from "@testing-library/react";
-import { SIGNUP_MOCK_ERROR, SIGNUP_MOCK_PAYLOAD, SIGNUP_MOCK_RESULT } from "../../utils/mocks";
+import { MOCK_PRESETS, SIGNUP_MOCK_ERROR, SIGNUP_MOCK_PAYLOAD, SIGNUP_MOCK_RESULT } from "../../utils/mocks";
 import "@types/jest";
 import "@testing-library/jest-dom";
 import "@testing-library/jest-dom/extend-expect";
 import user from "@testing-library/user-event";
 import { act } from "react-dom/test-utils";
-import { TestService } from "../../utils/TestServiceClass";
-const tapi = new TestService("alive");
 
 const store = createStore(
   allReducers,
@@ -31,134 +29,6 @@ window.HTMLMediaElement.prototype.pause = () => { /* do nothing */ };
 // @ts-ignore
 window.HTMLMediaElement.prototype.addTextTrack = () => { /* do nothing */ };
 
-//TODO IMPLEMENT WINDOW NAVIGATION window.location.assign() for signup test
-
-// const delay = (ms: number): Promise<void> => new Promise((resolve) => setTimeout(() => resolve(), ms));
-
-
-
-describe("Test rendering signup correctly", () => {
-
-  const originalFetch = global.fetch;
-
-  afterEach(() => {
-    cleanup();
-    global.fetch = originalFetch;
-    localStorage.clear();
-  });
-
-  it("Render the home page and then click sign up button to go to that page", async () => {
-
-    expect(tapi.alive()).toBe("alive");
-    const history = createMemoryHistory();
-
-    render(
-      <Provider store={store}>
-        <Router history={history}>
-          <App />
-        </Router>
-      </Provider>
-    );
-
-    const link = await screen.findByText(/^Sign\sUp$/g);
-    expect(link).toBeInTheDocument();
-    fireEvent.click(link);
-
-    const formEls = {
-      username: screen.getByPlaceholderText(/my_username/g),
-      email: screen.getByPlaceholderText(/example@email.com/g),
-      password: screen.getByPlaceholderText(/\*\*\*\*\*\*\*\*\*\*\*\*\*\*\*/g),
-      signup: screen.getAllByRole("button", { name: "Sign Up" }).find((btn) => {
-        return btn.classList.contains("form-btn");
-      }) as HTMLElement
-    };
-
-    expect(formEls.username).toBeInTheDocument();
-    expect(formEls.email).toBeInTheDocument();
-    expect(formEls.password).toBeInTheDocument();
-    expect(formEls.signup).toBeInTheDocument();
-
-  });
-});
-
-describe("test signup functionality", () => {
-
-  const originalFetch = global.fetch;
-  
-  afterEach(() => {
-    cleanup();
-    global.fetch = originalFetch;
-    localStorage.clear();
-  });
-
-  it("Checks the input fields are available and can submit with a stubbed api", async () => {
-    // @ts-ignore trying to mock fetch
-    global.fetch = jest.fn(() =>
-      Promise.resolve({
-        json: () => Promise.resolve(SIGNUP_MOCK_RESULT),
-      })
-    );
-    const history = createMemoryHistory();
-    render(
-      <Provider store={store}>
-        <Router history={history}>
-
-        </Router>
-          <App />
-      </Provider>
-    );
-
-    const page = (await screen.findAllByRole("link", { name: "Sign Up" })).find((el) => {
-      return el.classList.contains("nav-button");
-    }) as HTMLElement;
-    expect(page).toBeInTheDocument();
-    fireEvent.click(page);
-
-    const formEls = {
-      username: screen.getByPlaceholderText(/my_username/g) as HTMLInputElement,
-      email: screen.getByPlaceholderText(/example@email.com/g) as HTMLInputElement,
-      password: screen.getByPlaceholderText(/\*\*\*\*\*\*\*\*\*\*\*\*\*\*\*/g) as HTMLInputElement,
-      btn: screen.getAllByRole("button", { name: "Sign Up" }).find((btn) => {
-        return btn.classList.contains("form-btn");
-      }) as HTMLElement
-    };
-
-    //check the input elements are in the document
-    expect(formEls.username).toBeInTheDocument();
-    expect(formEls.email).toBeInTheDocument();
-    expect(formEls.password).toBeInTheDocument();
-    expect(formEls.btn).toBeInTheDocument();
-
-    // type inputs to form fields and submit
-
-    user.type(formEls.username, SIGNUP_MOCK_PAYLOAD.username);
-    user.type(formEls.email, SIGNUP_MOCK_PAYLOAD.email);
-    user.type(formEls.password, SIGNUP_MOCK_PAYLOAD.password);
-    
-    user.clear(formEls.username);
-    user.clear(formEls.email);
-    user.clear(formEls.password);
-    expect(formEls.username.value).toBe("");
-    expect(formEls.email.value).toBe("");
-    expect(formEls.password.value).toBe("");
-    
-    user.type(formEls.username, SIGNUP_MOCK_PAYLOAD.username);
-    user.type(formEls.email, SIGNUP_MOCK_PAYLOAD.email);
-    user.type(formEls.password, SIGNUP_MOCK_PAYLOAD.password);
-    expect(formEls.username.value).toBe(SIGNUP_MOCK_PAYLOAD.username);
-    expect(formEls.email.value).toBe(SIGNUP_MOCK_PAYLOAD.email);
-    expect(formEls.password.value).toBe(SIGNUP_MOCK_PAYLOAD.password);
-
-    //click signup to simulate api mock
-    
-    await act(async () => {
-      formEls.btn.dispatchEvent(new MouseEvent("click", { bubbles: true }));
-    });
-
-    expect(fetch).toHaveBeenCalledTimes(2);
-
-  });
-});
 describe("Tests network error message", () => {
 
   const originalFetch = global.fetch;
@@ -232,13 +102,16 @@ describe("need to impl window.location.assign()", () => {
   });
 
   it("app tries to redirect after successful signup", async () => {
+    const fakeFetchRes = (value: any): Promise<{ status: 200, json: () => 
+      Promise<any>; }> => Promise.resolve({ status: 200, json: () => Promise.resolve(value)});
+    const mockFetch = jest.fn()
+                      //default
+                      // .mockReturnValue("kdfjkdj")
+                      .mockReturnValueOnce(fakeFetchRes(SIGNUP_MOCK_RESULT))
+                      .mockReturnValueOnce(fakeFetchRes({ preset: { presetName: "waves" } }))
+                      .mockReturnValueOnce(fakeFetchRes({ presets: MOCK_PRESETS }));
     // @ts-ignore
-    global.fetch = jest.fn(() =>
-      Promise.resolve({
-        status: 200,
-        json: () => Promise.resolve(SIGNUP_MOCK_RESULT),
-      })
-    );
+    global.fetch = mockFetch;
     const history = createMemoryHistory();
 
     render(
@@ -300,6 +173,7 @@ describe("need to impl window.location.assign()", () => {
     });
 
     expect(fetch).toHaveBeenCalledTimes(2);
+    // expect(fetch).toHaveBeenNthCalledWith(2, "dkjfkdfj");
 
     expect(screen.getByTestId("location-display")).toHaveTextContent("/");
