+++ client/src/components/Modal/DeletePresetConfirmModal.tsx
@@ -1,25 +1,31 @@
-import React, { useState } from "react";
+import React from "react";
 import { escape } from "he";
 import API from "../../utils/ApiService";
 import { AuthService as Auth } from "../../utils/AuthService";
 import { useDispatch, useSelector } from "react-redux";
 import { MyRootState } from "../../types";
+import { deletePreset } from "../../actions/preset-button-actions";
 interface DeletePresetConfirmModalProps {
+  context: { btnId: string; };
   onConfirm: React.MouseEventHandler<HTMLElement>;
   onCancel: React.MouseEventHandler<HTMLElement>;
 }
 
-const SavePresetModal: React.FC<DeletePresetConfirmModalProps> = ({ 
+const DeletePresetModal: React.FC<DeletePresetConfirmModalProps> = ({ 
   onConfirm,
   onCancel,
+  context: { btnId }
 }) => {
-  const [ error, setError ] = useState<string>("");
+  // const [ error, setError ] = useState<string>("");
   const { presetButtons } = useSelector((state: MyRootState) => state.presetButtonsListState);
   const dispatch = useDispatch();
 
   // TODO: implement the api service for calling the api 
   // to delete the preset we confirm on the modal for a particular preset
   // that we clicked while delete mode was on
+  async function deleteThePreset(): Promise<void> {
+    await API.deletePreset(btnId, Auth.getToken() as string);
+  };
 
   
   return (
@@ -66,7 +72,8 @@ const SavePresetModal: React.FC<DeletePresetConfirmModalProps> = ({
       `)}}>
 
       </style>
-      <div 
+      <div
+        data-testid="delete-modal"
         style={{ 
           display: "flex", 
           flexDirection: "column", 
@@ -90,6 +97,8 @@ const SavePresetModal: React.FC<DeletePresetConfirmModalProps> = ({
           className="modal-confirm-button"
           onClick={(event) => {
             onConfirm(event); //closes modal and deletes preset from user's collection
+            dispatch(deletePreset(presetButtons, btnId));
+            void deleteThePreset();
           }}
         >
           YES
@@ -104,4 +113,4 @@ const SavePresetModal: React.FC<DeletePresetConfirmModalProps> = ({
 };
 
 
-export default SavePresetModal;
+export default DeletePresetModal;
