+++ client/src/jest.config.ts
@@ -4,6 +4,9 @@ export async function MyConfig(): Promise<Config.InitialOptions> {
   return {
     verbose: true,
     watchAll: true,
+    // restoreMocks: true, // even with this, jest mocks are still leaking into other tests!!!!
+    // clearMocks: true,
+    // resetMocks: true,
     modulePaths: ["<rootDir>"],
     transformIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/**/*.js"],
     testMatch: ["**/?(*.)+(spec|test).tsx"],
