+++ client/src/reducers/deleteModalReducer.ts
@@ -3,7 +3,8 @@ import { IDeleteModalState, IDeleteModalAction } from "../types";
 
 const modalReducer = (
   state: IDeleteModalState = {
-    deleteModalIsOpen: false
+    deleteModalIsOpen: false,
+    deleteModalContext: { btnId: "" }
   }, 
   action: IDeleteModalAction
 ): IDeleteModalState => {
@@ -12,6 +13,11 @@ const modalReducer = (
       return {
         ...state,
         deleteModalIsOpen: action.payload as boolean
+      };
+    case "SET_DELETE_MODAL_CONTEXT":
+      return {
+        ...state,
+        deleteModalContext: action.payload as { btnId: string; }
       };
     default: return state;
   }
