+++ client/src/utils/mocks.ts
@@ -4,7 +4,10 @@ import jwt from "jsonwebtoken";
 const uuid = require("uuid");
 
 export const SAVE_DEFAULT_MOCK_SUCCESS = {
-  updated: "presetName goes here hahahah"
+  preset: {
+    presetName: "new preset yo",
+    animVarCoeff: "74",
+  },
 };
 export const SAVE_DEFAULT_MOCK_ERROR = {
   error: "ERROR"
@@ -15,12 +18,15 @@ export const MOCK_SIGN_TOKEN_ARGS = {
   uuid: uuid.v4(),
   _id: Math.random() * 1293 + "yo what up",
 };
-export const MOCK_PRESETS = [{presetName: "rainbowTest", animVarCoeff: "64", _id: "6200149468fe291e26584e4b"},
-{presetName: "v2", animVarCoeff: "64", _id: "6200149468fe291e26584e4c"},
-{presetName: "waves", animVarCoeff: "64", _id: "6200149468fe291e26584e4d"},
-{presetName: "spiral", animVarCoeff: "64", _id: "6200149468fe291e26584e4e"},
-{presetName: "fourSpirals", animVarCoeff: "64", _id: "6200149468fe291e26584e4f"},
-{presetName: "dm5", animVarCoeff: "64", _id: "6200149468fe291e26584e50"}];
+export const MOCK_PRESETS = [
+  {presetName: "rainbowTest", animVarCoeff: "64", _id: "6200149468fe291e26584e4b"},
+  {presetName: "v2", animVarCoeff: "64", _id: "6200149468fe291e26584e4c"},
+  {presetName: "waves", animVarCoeff: "64", _id: "6200149468fe291e26584e4d"},
+  {presetName: "spiral", animVarCoeff: "64", _id: "6200149468fe291e26584e4e"},
+  {presetName: "fourSpirals", animVarCoeff: "64", _id: "6200149468fe291e26584e4f"},
+  {presetName: "dm5", animVarCoeff: "64", _id: "6200149468fe291e26584e50"},
+  {presetName: "bogus", animVarCoeff: "64", _id: "6200149468fe291e26584e51"},
+];
 
 export const ASSERT_ANIMATION = {
   clearLed: "led1-1",
@@ -137,7 +143,10 @@ export const SIGNUP_MOCK_PAYLOAD = {
 };
 
 export const SIGNUP_MOCK_RESULT = {
-  token: "heres a token",
+  token: jwt.sign({
+    username: "weeeee",
+    _id: "weeeeeee",
+  }, "SECRET!!!!", { expiresIn: "1h" }),
   _id: "heres an id"
 };
