+++ server/controllers/UserController.ts
@@ -47,17 +47,15 @@ export const UserController = {
   ): Promise<Response | void> {
     try {
       const { _id } = req.body;
-      const user = await User.findOneAndUpdate(
+      await User.findOneAndUpdate(
         { email: req!.user!.email },
         {
           $pull: {
             presets: { _id },
           },
-        },
-        { new: true }
+        }
       );
-
-      res.status(200).json({ presets: user!.presets });
+      res.status(200).json({ message: "deleted the preset" });
     } catch (error) {
       console.error(error);
       res.status(500).json({ error });
